- more SimulationRuntime/c
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@10386 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/meta/meta_modelica_builtin.c
@@ -711,6 +711,6 @@ modelica_real realMaxLit()
 
 modelica_integer intMaxLit()
 {
-  return LONG_MAX / 2;
+  return INT_MAX / 2;
 }
